Chat duplicate detection code is confused by the same response for different referenced messages

+++ scripts/cron/chatdups.php
@@ -15,13 +15,16 @@ foreach ($chats as $chat) {
     $chatmsgs = $dbhr->preQuery("SELECT * FROM chat_messages WHERE date > ? AND chatid = ?;", [ $mysqltime, $chat['id'] ]);
 
     $lastmsg = NULL;
+    $lastref = NULL;
     $lastid = NULL;
+
     foreach ($chatmsgs as $msg) {
-        if ($lastmsg && $lastmsg == $msg['message']) {
+        if ($lastmsg && $lastmsg == $msg['message'] && $lastref == $msg['refmsgid']) {
             error_log("{$chat['id']} $lastid and {$msg['id']}");
             $dbhm->preExec("DELETE FROM chat_messages WHERE id = ?;", [ $msg['id'] ]);
         } else {
             $lastmsg = $msg['message'];
+            $lastref = $msg['refmsgid'];
             $lastid = $msg['id'];
         }
     }
